Move 'VersionOrPath' to 'selected.rs', canonicalize the path
